requirements.txt
# --- Core ---
python-dotenv>=1.0
pydantic>=2
requests>=2.32
tqdm>=4.66
numpy>=1.26.0
fastapi>=0.100.0
uvicorn>=0.23.0

# --- Vector Store ---
pymilvus>=2.4.6

# --- LLM / OpenAI ---
openai>=1.3.0
langchain>=0.2.16
langchain-community>=0.2.7
langchain-openai>=0.1.23
langchain-text-splitters>=0.2.2
langsmith>=0.1.17

# --- RAG / Parsing ---
beautifulsoup4>=4.12.3
lxml>=5.2.2
readability-lxml>=0.8.1
rank-bm25>=0.2.2

# --- Vector DB (Milvus) ---
pymilvus>=2.4.6

# --- Guardrails ---
guardrails-ai>=0.5.0

# --- Eval ---
datasets>=2.20.0

# Eğer OpenAI embedding kullanıyorsan BU KADAR yeter.
# Yerel embedding istersen (ağırdır, M1’de torch indirir):
# sentence-transformers>=2.6.1
# torch>=2.3.0

# HTTP & HTML
requests>=2.32.0
beautifulsoup4>=4.12.3
lxml>=5.2.2 

# Sayısal
numpy>=1.26.0 

# OpenAI API (yeni client: OpenAI())
openai>=1.40.0

# Milvus client
pymilvus==2.6.2
grpcio==1.67.1

# Ayar
pydantic>=2.8.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.1
urllib3>=1.26.18,<2.4.0
kubernetes==34.1.0

